+++ app/Http/Controllers/ForumController.php
@@ -8,7 +8,9 @@ use App\Models\Comment;
 use App\Models\CommentLike;
 use App\Models\Forum;
 use App\Models\ForumLike;
+use App\Models\ViewLog;
 use Auth;
+use Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Str;
@@ -19,6 +21,16 @@ class ForumController extends Controller
     {
         $post = Forum::find($post->id);
         $comments = Comment::where('forum_id', $post->id)->orderBy('created_at', 'desc')->get();
+
+        // View Log
+        if (Auth::check()){
+            $user = Auth::user();
+            $log = ViewLog::where('user_id', $user->id)->where('forum_id', $post->id)->first();
+            if ($log != null){
+                $user->viewLogs()->detach($post);
+            }
+            $user->viewLogs()->attach($post, ['created_at' => now()]);
+        }
         return view('post-detail', ['post' => $post, 'comments' => $comments, 'filter'=> "Newest"]);
     }
 
@@ -39,10 +51,13 @@ class ForumController extends Controller
         }
         return view('post-detail', ['post' => $post, 'comments' => $comments, 'filter'=>$filter]);
     }
-    
+
     public function getEditPostPage(Forum $post)
     {
         $post = Forum::find($post->id);
+        if (! Gate::allows('admin-access') && $post->user->id != Auth::id()) {
+            abort(403);
+        }
         $categories = Category::all();
         $categoryTypes = CategoryType::all();
         return view('edit-post', ['post' => $post, 'categories' => $categories, 'categoryTypes' => $categoryTypes]);
@@ -51,17 +66,25 @@ class ForumController extends Controller
     public function editPost(Request $request, Forum $post)
     {
         $validated = $request->validate([
-            'postTitle' => 'required',
+            'postTitle' => 'required|min:5',
             'programmingLanguage' => 'required',
             'postType' => 'required',
             'image' => 'image|nullable',
-            'content' => 'required',
+            'content' => 'required|min:5|max:1000',
         ]);
         $post = Forum::find($post->id);
         $post->title = $validated["postTitle"];
         $post->content = $validated["content"];
         $post->category_id = $validated['programmingLanguage'];
         $post->category_type_id = $validated['postType'];
+        if ($request->has('image')){
+            $path = $request->file('image')->store('images', 'public');
+            $validated['image'] = $path;
+            if ($post->image !== null){
+                Storage::disk('public')->delete($post->image);
+            }
+            $post->image = $validated['image'];
+        }
         $post->save();
 
         $comments = Comment::where('forum_id', $post->id)->get();
@@ -72,6 +95,9 @@ class ForumController extends Controller
 
     public function deletePost(Forum $post){
         $post = Forum::find($post->id);
+        if (! Gate::allows('admin-access') && $post->user->id != Auth::id()) {
+            abort(403);
+        }
         if ($post->image !== null){
             Storage::disk('public')->delete($post->image);
         }
@@ -82,7 +108,17 @@ class ForumController extends Controller
     public function getAllPostPage()
     {
         $posts = Forum::orderby('created_at', 'desc')->paginate(10);
-        return view('posts', ['posts' => $posts,'programmingLanguage' => "All", 'postType' => "All", 'sortBy' => "New", 'search' => ""]);
+
+        $logs = collect();
+        if (Auth::check()){
+            $user = Auth::user();
+            $listPosts = ViewLog::where('user_id', $user->id)->orderBy('created_at', 'desc')->take(10)->get('forum_id');
+            foreach($listPosts as $postId){
+                $post = Forum::find($postId)->first();
+                $logs->push($post);
+            }
+        }
+        return view('posts', ['posts' => $posts,'programmingLanguage' => "All", 'postType' => "All", 'sortBy' => "New", 'search' => "", 'logs' => $logs]);
     }
 
     public function getCreatePostPage()
@@ -98,13 +134,12 @@ class ForumController extends Controller
     public function createPost(Request $request)
     {
         $validated = $request->validate([
-            'postTitle' => 'required',
+            'postTitle' => 'required|min:5',
             'programmingLanguage' => 'required',
             'postType' => 'required',
             'image' => 'image|nullable',
-            'content' => 'required',
+            'content' => 'required|min:5|max:1000',
         ]);
-
         $post = new Forum();
         $post->user_id = Auth::user()->id;
         $post->category_id = $validated['programmingLanguage'];
@@ -117,6 +152,10 @@ class ForumController extends Controller
             $post->image = $validated['image'];
         }
         $post->save();
+
+        // Badges
+        $badgeController = new BadgeController();
+        $badgeController->addBadge('create_forum');
 
         $comments = Comment::where('forum_id', $post->id)->get();
         return redirect()
@@ -153,7 +192,7 @@ class ForumController extends Controller
     public function filterPosts(string $programmingLanguage, string $postType, string $sortBy, string $search = ""){
 
         $posts = Forum::all();
-        
+
         if (Str::upper($programmingLanguage) !== "ALL"){
             $posts = $posts->filter(function($post) use($programmingLanguage) {
                 return $post->category->category_name == $programmingLanguage;
@@ -170,7 +209,7 @@ class ForumController extends Controller
         foreach($posts as $post){
             array_push($ids, $post->id);
         }
-        
+
         if (Str::upper($sortBy) === "NEW"){
             $posts = Forum::whereIn('id', $ids)->orderby('created_at', 'desc')->where('title', 'LIKE', '%'. $search.'%')->paginate(6);
         }
@@ -184,6 +223,21 @@ class ForumController extends Controller
             $posts = Forum::whereIn('id', $ids)->withCount('userLikes')->where('title', 'LIKE', '%'. $search.'%')->orderby('user_likes_count', 'asc')->paginate(6);
         }
 
-        return view('posts', ['posts' => $posts ,'programmingLanguage' => $programmingLanguage, 'postType' => $postType, 'sortBy' => $sortBy, 'search' => $search]);
+        $logs = collect();
+        if (Auth::check()){
+            $user = Auth::user();
+            $listPosts = ViewLog::where('user_id', $user->id)->orderBy('created_at', 'desc')->take(10)->get('forum_id');
+            foreach($listPosts as $postId){
+                $post = Forum::find($postId)->first();
+                $logs->push($post);
+            }
+        }
+        return view('posts', ['posts' => $posts ,'programmingLanguage' => $programmingLanguage, 'postType' => $postType, 'sortBy' => $sortBy, 'search' => $search, 'logs'=> $logs]);
+    }
+
+    public function clearLogs(){
+        $user = Auth::user();
+        ViewLog::where('user_id', $user->id)->delete();
+        return redirect()->route('posts');
     }
 }
