+++ app/Http/Controllers/ReplyController.php
@@ -4,15 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Models\Reply;
 use Auth;
+use Gate;
 use Illuminate\Http\Request;
 
 class ReplyController extends Controller
 {
     public function createReply(string $commentId, Request $request){
+        $validated = $request->validate([
+            'content' => 'required|max:100',
+        ]);
         $reply = new Reply();
         $reply->user_id = auth()->user()->id;
         $reply->comment_id = $commentId;
-        $reply->content = $request['content'];
+        $reply->content = $validated['content'];
         $reply->save();
         return response()->json([
             'success' => true,
@@ -23,6 +27,9 @@ class ReplyController extends Controller
 
     public function deleteReply(string $replyId){
         $reply = Reply::find($replyId);
+        if (! Gate::allows('admin-access') && $reply->user->id != Auth::id()) {
+            abort(403);
+        }
         $reply->delete();
         return back()->with('success','Reply Deleted');
     }
@@ -50,6 +57,22 @@ class ReplyController extends Controller
         return response()->json([
             'success' => true,
             'likes' => $likes
+        ]);
+    }
+
+    public function editReply(string $reply, Request $request)
+    {
+        $validated = $request->validate([
+            'content' => 'max:200',
+        ]);
+        $reply = Reply::find($reply);
+        $reply->content = $validated['content'];
+        $reply->updated_at = now();
+        $reply->save();
+        
+        return response()->json([
+            'success' => true,
+            'replyId' => $reply->id
         ]);
     }
 }
